backend/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy
psycopg2-binary
Werkzeug==2.2.2
flask-cors
Pillow
flask-bcrypt
python-dotenv
flask-session
opencv-python-headless
pytesseract
deepl
requests
langdetect
Flask-Migrate==3.1.0
doctr
python-doctr[tf]
tensorflow
numpy